requirements.txt
Pillow>=10.0.0
reportlab>=4.0.0
rembg>=2.0.0
onnxruntime>=1.15.0 